requirements.txt
aiogram==2.22.2
requests~=2.28.1
python-dotenv~=0.21.0
loguru~=0.6.0